Fixes for templated config rendering
There were a few things needed to make the previous commits fully
functional. First, the images were not wired in to the bootstrap
manifest rendering. Second, the templates themselves had a few
issues that needed to be resolved (capitalization, indentation, etc.).

This passes the necessary image details into the bootstrap manifest
rendering code and cleans up the templates. After this patch the
bootstrap services can be started successfully.