Create unique connections for NICs
The default NetworkManager configuration uses multi-connect=3, meaning that
default_connection as a profile can be auto and manually activated multiple
times. Usually, this isn't a problem, but it seems that nmstate currently
can't handle this. A bug is filed [1], but this works around that by moving
each of the active connections to their own profile.

1. https://github.com/nmstate/nmstate/issues/1474

Signed-off-by: Brad P. Crochet <brad@redhat.com>